requirements.txt
PyJWT==1.1
PySerial==2.7
